Refactor selection sort implementation to handle floating point values and update related functions

+++ 1st_Year/2SEM/DS/a1q3_selection_sort.c
@@ -4,7 +4,7 @@
  * Date: 24 March, 2025
  *
  * Program: Selection Sort Implementation
- * Description: This program sorts an array of integers in ascending order using the selection sort algorithm.
+ * Description: This program sorts an array of floating point values in ascending order using the selection sort algorithm.
  *
  * Author: Biranchi Kulesika
  * Date: 25 March, 2025
@@ -14,7 +14,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-void selectionSort(int *arr, int arrSize)
+void selectionSort(float *arr, int arrSize)
 {
 	for (int i = 0; i < arrSize - 1; i++)
 	{
@@ -32,11 +32,11 @@ void selectionSort(int *arr, int arrSize)
 	}
 }
 
-void printArray(int *arr, int arrSize)
+void printArray(float *arr, int arrSize)
 {
 	for (int i = 0; i < arrSize; i++)
 	{
-		printf("%d ", arr[i]);
+		printf("%.2f ", arr[i]);
 	}
 	printf("\n");
 }
@@ -54,7 +54,7 @@ int main()
 		return 1;
 	}
 
-	int *arr = (int *)malloc(n * sizeof(int));
+	float *arr = (float *)malloc(n * sizeof(float));
 	if (!arr)
 	{
 		printf("Memory allocation failed!\n");
@@ -64,7 +64,7 @@ int main()
 	printf("Enter %d elements separated by space:\n", n);
 	for (int i = 0; i < n; i++)
 	{
-		scanf("%d", &arr[i]);
+		scanf("%f", &arr[i]);
 	}
 
 	printf("The entered array is:\n");
